+++ source-code/sdk-js/build.js
@@ -11,8 +11,8 @@ const options = {
 	platform: "node",
 	format: "esm",
 	target: "esnext",
-	sourcemap: !dev,
-	minify: !dev,
+	sourcemap: true,
+	minify: false,
 	splitting: false,
 	external: ["svelte", "@sveltejs/kit", "typescript", "ts-morph"],
 	plugins: [