+++ source-code/sdk-js/src/adapter-sveltekit/ast-transforms/+layout.svelte.ts
@@ -32,6 +32,7 @@ export const transformLayoutSvelte = (
 const transformScript = (filePath: string, config: TransformConfig, code: string) => {
 	const sourceFile = codeToSourceFile(code, filePath)
 
+	addImport(sourceFile, "@inlang/sdk-js/adapter-sveltekit/client/shared", "addRuntimeToGlobalThis")
 	addImport(sourceFile, "@inlang/sdk-js/adapter-sveltekit/shared", "getRuntimeFromData")
 	addImport(
 		sourceFile,
@@ -48,21 +49,26 @@ const transformScript = (filePath: string, config: TransformConfig, code: string
 
 	const index = addDataExportIfMissingAndReturnInsertionIndex(sourceFile)
 
+	// TODO: add `addRuntimeToGlobalThis` code only if needed
 	sourceFile.insertStatements(
 		index + 1,
 		dedent`
-		$: if (browser) {
-			addRuntimeToContext(getRuntimeFromData(data))
-			;({ i, languageTag } = getRuntimeFromContext())
-		}
-	`,
+			$: if (browser) {
+				addRuntimeToGlobalThis(getRuntimeFromData(data))
+				addRuntimeToContext(getRuntimeFromData(data))
+				;({ i, languageTag } = getRuntimeFromContext())
+				document.body.parentElement?.setAttribute('lang', language)
+			}
+		`,
 	)
+
 	sourceFile.insertStatements(
 		index + 1,
 		dedent`
-		addRuntimeToContext(getRuntimeFromData(data))
-		let { i, languageTag } = getRuntimeFromContext()
-	`,
+			addRuntimeToGlobalThis(getRuntimeFromData(data))
+			addRuntimeToContext(getRuntimeFromData(data))
+			let { i, languageTag } = getRuntimeFromContext()
+		`,
 	)
 
 	return nodeToCode(sourceFile)