+++ source-code/sdk-js/src/adapter-sveltekit/examples/test-app/src/routes/[lang]/+page.svelte
@@ -1,13 +1,15 @@
 <script>
 	import { i, languageTags, switchLanguage } from "@inlang/sdk-js"
+	import { onMount } from "svelte"
+	import { test } from "./test.js"
 
-	console.log(11, switchLanguage);
-
-	export let data
+	onMount(() => {
+		test()
+	})
 </script>
 
 {#each languageTags as lang}
-	<button on:click={() => switchLanguage(lang)}>{lang}</button>
+	<button on:click={() =>switchLanguage(lang)}>{lang}</button>
 {/each}
 
 <h1>{i("welcome")}</h1>