+++ source-code/sdk-js/src/adapter-sveltekit/runtime/client/reactive-workaround/context.ts
@@ -1,11 +1,12 @@
 // workaround for https://github.com/inlang/inlang/issues/1091
-
-import { setContext } from "svelte"
+import type * as Ast from "@inlang/core/ast"
 import { get } from "svelte/store"
 import type { RelativeUrl } from "../../../../index.js"
-import { inlangSymbol } from "../../shared/utils.js"
 import type { SvelteKitClientRuntime } from "../index.js"
-import { getRuntimeFromContext as getRuntimeFromContextShared } from "../shared/context.js"
+import {
+	getRuntimeFromContext as getRuntimeFromContextShared,
+	addRuntimeToContext as addRuntimeToContextShared,
+} from "../shared/context.js"
 import type * as Runtime from "../../../../runtime/index.js"
 import { goto } from "$app/navigation"
 import { page } from "$app/stores"