+++ source-code/website/src/pages/documentation/index.page.server.tsx
@@ -2,7 +2,7 @@ import type { OnBeforeRender } from "@src/renderer/types.js"
 import type { PageProps } from "./index.page.jsx"
 import { tableOfContents, FrontmatterSchema } from "../../../../../documentation/tableOfContents.js"
 import { parseMarkdown } from "@src/services/markdown/index.js"
-import { RenderErrorPage } from "vite-plugin-ssr/RenderErrorPage"
+import { RenderErrorPage } from "vite-plugin-ssr/server"
 
 /**
  * the table of contents without the html for each document
@@ -33,11 +33,9 @@ await generateIndexAndTableOfContents()
 
 // should only run server side
 export const onBeforeRender: OnBeforeRender<PageProps> = async (pageContext) => {
-	let headings = []
 	// dirty way to get reload of markdown (not hot reload though)
 	if (import.meta.env.DEV) {
 		await generateIndexAndTableOfContents()
-		headings = await generateHeadings(pageContext.urlPathname)
 	}
 	if (!Object.keys(index).includes(pageContext.urlPathname)) {
 		throw RenderErrorPage({ pageContext: { is404: true } })
@@ -45,9 +43,8 @@ export const onBeforeRender: OnBeforeRender<PageProps> = async (pageContext) => 
 	return {
 		pageContext: {
 			pageProps: {
-				markdown: index[pageContext.urlPathname]!,
+				markdown: index[pageContext.urlPathname],
 				processedTableOfContents: processedTableOfContents,
-				headings: headings,
 			},
 		},
 	}
@@ -61,7 +58,7 @@ async function generateIndexAndTableOfContents() {
 		const frontmatters: { frontmatter: any }[] = []
 		for (const document of documents) {
 			const markdown = parseMarkdown({
-				text: document,
+				text: document.import,
 				FrontmatterSchema,
 			})
 			// not pushing to processedTableOfContents directly in case
@@ -71,26 +68,4 @@ async function generateIndexAndTableOfContents() {
 		}
 		processedTableOfContents[category] = frontmatters
 	}
-}
-
-/**
- * Generates the headings
- */
-async function generateHeadings(urlPathname: string) {
-	const headings: PageProps["headings"] = []
-
-	const markdown = index[urlPathname]
-	if (markdown && markdown.renderableTree) {
-		for (const heading of markdown.renderableTree.children) {
-			if (heading.name === "Heading") {
-				if (heading.children[0].name) {
-					headings.push(heading.children[0])
-				} else {
-					headings.push(heading)
-				}
-			}
-		}
-	}
-
-	return headings
 }