+++ source-code/website/src/pages/editor/@host/@owner/@repository/Layout.tsx
@@ -214,7 +214,9 @@ export function Layout(props: { children: JSXElement }) {
 							>
 								<sl-dropdown prop:distance={8}>
 									<sl-button prop:size="small" slot="trigger">
-										<IconAdd slot="prefix" />
+										<button slot="prefix">
+											<IconAdd class="w-4 h-4" />
+										</button>
 										Filter
 									</sl-button>
 									<sl-menu>
@@ -425,8 +427,8 @@ function LanguageFilter(props: { clearFunction: any }) {
 						is
 					</p>
 				</div>
-				<button slot="clear-icon">
-					<IconClose />
+				<button slot="clear-icon" class="p-0.5">
+					<IconClose class="w-4 h-4" />
 				</button>
 
 				<div class="flex px-3 gap-2 text-sm font-medium">
@@ -518,8 +520,8 @@ function LintFilter(props: { clearFunction: any }) {
 					</button>
 				</Show>
 			</div>
-			<button slot="clear-icon">
-				<IconClose />
+			<button slot="clear-icon" class="p-0.5">
+				<IconClose class="w-4 h-4" />
 			</button>
 
 			<div class="flex px-3 gap-2 text-sm font-medium">