+++ source-code/website/src/pages/editor/@host/@owner/@repository/components/PatternEditorTipTap.tsx
@@ -26,7 +26,7 @@ export const TipTapEditor = (props: {
 	createEffect(() => {
 		if (editor) {
 			const json = useEditorJSON(() => editor())
-			console.log(json())
+			console.debug(json())
 		}
 	})